Fix non-working successful response assertion
So the previous version would still pass if the controller was changed
to return 500. And this was because these are two different it-blocks
that wouldn't remember what the other might have had as responsed. I can
only presume that a response object will always return successful? true
regardless if a request has been sent or not yet.